+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [PEP 440](https://www.python.org/dev/peps/pep-0440/)
 and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Added
+* The ability to specify swaths and minimum number of bursts when using tool.
+
+### Changed
+* Moved all search/download functionality to `search.py` module.
+
 ## [0.3.5]
 
 ### Fixed